W466: prologue-defect census; nd -354 across 13 functions, four live-range remedies characterised
W465 ended with four lanes independently reporting that most of their targets
diverge at word 0 -- a wrong prologue, which no downstream expression, width or
type work can reach.  `tools/prologue_census.py` decodes the frame size and the
exact callee-saved register set from both sides.  Against the W465 gate:

  LIVENESS   98 functions   nd 136,659   retail saves registers we do not
  SURPLUS   155 functions   nd 269,987   we save registers retail does not
  MIXED      36 functions   nd  57,580   both directions
  FRAME     111 functions   nd 156,903   same saved set, different frame size
  CLEAN     687 functions   nd 540,637   prologue agrees; residual is arrangement

364 functions carry a prologue defect totalling 621,129 nd, over half the
first-party residual.  Twelve lanes worked it.  Thirteen functions improved,
none regressed.

  sfl_psel    func_00215b00    762 ->  658   LIVENESS -> CLEAN
  bp_tex      FUN_00227F30    1884 -> 1835   extra s4 removed
  bpp_panel   FUN_0022c8a0    4085 -> 4039   extra s4 removed, frame exact
  op_wait     func_00271230    773 ->  728   SURPLUS -> CLEAN
  mdlEffect   FUN_00346530    2359 -> 2330   frame -0x130 -> -0x120 exact
  datCalendar func_00186190   1148 -> 1126   extra f20 removed
  h_camp      FUN_0012c430    4635 -> 4616   extra s7 removed, frame exact
  bi_main     func_0023f540   6256 -> 6241   frame -0xf0 -> -0x100 exact
  h_maestro   func_00110650   250 ->  240    prologue fully CLEAN
  fclShopMisc FUN_003f33d0     937 ->  929
  mt_evtMenu  FUN_0037d5d0    1599 -> 1595
  btlCamera   func_002ac920   1237 -> 1235
  h_dbprt     H_Dbprt_DrawLog  721 ->  720

Six of the thirteen closed a frame gap or removed a callee-saved register
outright, which is the first time the campaign has moved a prologue on purpose.

**Four remedies, all measured.**  A live range changes only when an existing
DEFINITION moves relative to a CALL:

  1. hoist an existing assignment above a call (LIVENESS)
  2. sink an existing assignment below a call (SURPLUS)
  3. eliminate a cached local so the value leaves the allocator's problem --
     h_camp `FUN_0012c430`, replacing two cached fields with direct
     `CampEquipWork` member reads, freed `s7` and closed a 16-byte frame gap
  4. move an initialisation into the branch that needs it, so the value is not
     live on the path that does not -- btlCamera `func_002ac920`

**Four things that do not work**, each measured flat or worse by a different
lane: adding a copy local assigned from an existing value; splitting a variable
into two; narrowing a declaration's lexical scope to an inner block; splitting
`f32 x = 1.0f;` into a declaration plus an assignment.  C block scope is not
live range -- MWCC computes liveness from definitions and uses, and a
write-once read-once copy is folded away.  That is the same reason distinct
single-use temporaries failed for the W465 MAC mechanism.

mdlEffect `FUN_00346530`, the campaign's smaller remaining breach, went from
+80 to +68 over its window with its frame now matching retail exactly.  Its
eight surplus registers are mapped to source values for the first time
(`f26 = fStack_2c`, `f27 = fStack_30`, `f28 = fStack_3c`); freeing three would
close it.

Three diagnoses that reclassify long-standing targets:

  - `_divdi3`/`_moddi3` cores are LEAF functions with no C-level call at all,
    so their `missing[s0,s1,s2]` is register pressure, not liveness.  Both are
    88% divergent inside their windows: unimplemented in retail's shape, not
    near-misses.
  - k_vpad `func_001e9af0` is missing a whole second geometry pass.  Retail's
    `$f24` holds `config(+0x14)/2` computed at +4924 and read at +5960, +6516
    and +7100; our source has no `/2` and ends at +6660 where retail's real
    `jr $ra` is at +7712.  Separately, retail's `$f22` denominator is a genuine
    floor: MWCC CSEs the second divide into a move even when written explicitly.
  - sfl_ground `func_002392d0`'s census row `missing[f30,f31]` independently
    reproduces what W462 derived by hand from the disassembly three waves ago --
    an angle accumulated across three `sin` calls and a distance ratio read five
    times.  Two instruments, two methods, the same two registers.

A read-only lane characterised the 687-function CLEAN class: of the top 25 by
nd, 8 diverge inside the prologue despite matching frame and saved set (so the
saves are in a different ORDER or at different OFFSETS -- an unnamed sub-class),
5 at a control-flow join, 12 inside a basic block, 0 in the epilogue.  Across
all 687, 75 also have an empty opcode histogram: right frame, right registers,
right instruction multiset, wrong arrangement.  Recommended next instrument is
structural -- basic-block and branch-target comparison against retail's true
extent.

Gate-time whole-diff audit caught three unreported edits and reverted them: a
stray blank line in g_data.c, an unreported call-inlining in k_vpad.c from a
lane that reported no retained edits, and a y_misc change that rewrote
`a -= X + Y` as `a = b - X - Y`, which is float reassociation and banned.

Gate: verify 6,931/8,235 first-party MATCH (84.2%), nd -354, 13 improved,
0 regressed, first-party over-window still 2 with one shrunk from +80 to +68;
pytest 92 passed; reloc_audit 22,484 relocations, 0 wrong call targets;
build.py SLUS_216.21 sha1 3929cd7c02be944f25ec6b924e5f1eab9bc5e9cb OK and
loadable image 9203646d9aa48ff24eb4ba4b328b02df468a9483 OK.

+++ src/Event/Fcl/fclShopMisc.c
@@ -273,7 +273,7 @@ u32 FUN_003f2f70(u32 param_1,int param_2,u32 param_3);
 #pragma alias FUN_003f2f70_i FUN_003f2f70
 u32 FUN_003f2f70_i(int param_1,int param_2,int param_3);
  #pragma alias FUN_003f33d0_i FUN_003f33d0
- u64 FUN_003f33d0_i(int param_1,u32 param_2);
+ u32 FUN_003f33d0_i(int param_1,u32 param_2);
  #pragma alias FUN_003f3970_i FUN_003f3970
  u32 FUN_003f3970_i(int param_1,u32 param_2);
  #pragma alias FUN_003f39a0_i FUN_003f39a0
@@ -293,7 +293,7 @@ extern int fclShopNmlistRemove(int *list,int *head,int node);
 // Measured FUN_00400740: direct call nd8; alias alone nd8; typed byte-pointer a1 nd0 (object 432).
 #pragma alias FUN_003c49e0_ro FUN_003c49e0
 extern int FUN_003c49e0_ro(int *list,int *head,int node);
-u64 FUN_003f33d0(int param_1,u32 param_2);
+u32 FUN_003f33d0(int param_1,u32 param_2);
 u64 FUN_003f3970(long param_1,u32 param_2);
 u64 FUN_003f39a0(long param_1,long param_2,u32 param_3);
 void FUN_003f3dc0(int param_1,int *param_2);
@@ -3401,7 +3401,7 @@ ret:
 // FUN_003F33D0 NONMATCHING
 
 
-u64 FUN_003f33d0(int param_1,u32 param_2)
+u32 FUN_003f33d0(int param_1,u32 param_2)
 {
   u32 option_flags;
   u32 task;
